+++ src/ScottPlot5/ScottPlot5 Cookbook/Recipes/Introduction/Styling.cs
@@ -9,10 +9,12 @@ internal class Styling : RecipePageBase
         PageDescription = "How to customize plots",
     };
 
-    internal class BackgroundColors : RecipeTestBase
+    internal class StyleClass : RecipeTestBase
     {
-        public override string Name => "Background Colors";
-        public override string Description => "Figure and data area background colors can be customized.";
+        public override string Name => "Style Helper Functions";
+        public override string Description => "Plots contain many objects which can be customized individually " +
+            "by assigining to their public properties, but helper methods exist in the Plot's Style object " +
+            "that make it easier to customzie many items at once using a simpler API.";
 
         [Test]
         public override void Recipe()
@@ -20,8 +22,15 @@ internal class Styling : RecipePageBase
             myPlot.Add.Signal(Generate.Sin(51));
             myPlot.Add.Signal(Generate.Cos(51));
 
-            myPlot.FigureBackground = Colors.DarkGray;
-            myPlot.DataBackground = Colors.LightGray;
+            // visible items have public properties that can be customized
+            myPlot.XAxis.Label.Text = "Horizontal Axis";
+            myPlot.YAxis.Label.Text = "Vertical Axis";
+            myPlot.Title.Label.Text = "Plot Title";
+
+            // the Style object contains helper methods to easily style many items at once
+            myPlot.Style.Background(figure: Color.FromHex("#07263b"), data: Color.FromHex("#0b3049"));
+            myPlot.Style.ColorAxes(Color.FromHex("#a0acb5"));
+            myPlot.Style.ColorGrids(Color.FromHex("#0e3d54"));
         }
     }