+++ src/ScottPlot5/ScottPlot5/TODO.cs
@@ -2,14 +2,16 @@
 
 internal static class TODO
 {
-    /* ### TASKS TO COMPLETE AFTER THE FIRST RELEASE ###
+    /* ### SCOSCOTT'S RUNNING LIST OF UPCOMING TASKS ##    
      * DateTime tick generator
      * Manual tick generator
-     * Marker class
-     * Style classes to group common configuration options: LineStyle and FontStyle
+     * Make more marker class
+     * LinePattern is not currently supported
      * Improve signal plot rendering when several repeated values produce a vertical height of zero
      * AutoScale() is called in the first render if SetAxisLimits() was used to set only a single axis value
      * Add offset X and offset Y support to signal plots
      * Multiplot (#1337)
+     * Add more data source classes
+     * Add XLabel(), YLabel(), and Title() shortcuts methods
      */
 }